back - 어드민툴 제어 기능 API 에서 권한 체크하도록 수정

+++ backend/api/access_level.js
@@ -1,10 +1,10 @@
 import { asyncHandler } from "../utils/asyncHandler.js"
 import { saveUserLog } from "../utils/user_log.js"
 import { LOG_ACT_TYPE, LOG_CONTENTS_TYPE } from "../constants/userLogType.js"
+import { isAdminAuthorized } from "../utils/adminPermision.js"
 
 import db from "../mysql/db.js"
 import express from 'express'
-
 
 
 const router = express.Router()
@@ -30,16 +30,18 @@ router.get('/access_level/:id', asyncHandler(async (req, res) => {
    })
 }))
 
+// TODO 아이디 변경은 못하게 막기 (프론트도 막기)
 router.put('/access_level/:id', asyncHandler(async (req, res) => {
     const { id } = req.params
-    const { setting, adminUserId } = req.body
+    const { setting, adminUserId, adminToken } = req.body
 
-    // TODO 아이디 변경은 못하게 막기 (프론트도 막기)
-    
+    if (!await isAdminAuthorized(adminUserId, adminToken)) {
+        return res.status(400).send({ message: `[UPDATE ERROR] no admin authorized` })
+    }
 
     const [updateRes] = await db.query('UPDATE user_info SET user_type = ? WHERE id = ?', [setting.state, id])
     if (updateRes.affectedRows === 0) {
-        return res.status(404).send({ message: `[UPDATE ERROR] user_info with id ${id}` })
+        return res.status(500).send({ message: `[UPDATE ERROR] user_info with id ${id}` })
     }
 
     await saveUserLog(adminUserId, LOG_CONTENTS_TYPE.WEAPON_ACCOUNT_MANAGEMENT, LOG_ACT_TYPE.UPDATE, id, setting.state)
@@ -48,15 +50,19 @@ router.put('/access_level/:id', asyncHandler(async (req, res) => {
 }))
 
 router.delete('/access_level/:id', asyncHandler(async (req, res) => {
-    const { id }          = req.params
-    const { adminUserId } = req.body
+    const { id } = req.params
+    const { adminUserId, adminToken } = req.body
 
     // TODO 해당 유저 관련 DB 데이터 삭제 필요
     // TODO comment 쪽 데이터는 유지하기 or 삭제 상태로 업데이트
 
+    if (!await isAdminAuthorized(adminUserId, adminToken)) {
+        return res.status(400).send({ message: `[DELETE ERROR] no admin authorized` })
+    }
+
     const [results] = await db.query('DELETE FROM user_info WHERE id = ?', [id])
     if (results.affectedRows === 0) {
-        return res.status(404).send({ message: `[DELETE ERROR] user_info with id ${id}` })
+        return res.status(500).send({ message: `[DELETE ERROR] user_info with id ${id}` })
     }
 
     await saveUserLog(adminUserId, LOG_CONTENTS_TYPE.WEAPON_ACCOUNT_MANAGEMENT, LOG_ACT_TYPE.DELETE, id)
